Update the Kim-1's mkcbmfs to use c1541.
