+++ src/arch/commodore/build.py
@@ -1,5 +1,5 @@
 from build.ab import simplerule, TargetsMap, filenameof, Rule, Target
-from tools.build import mkcpmfs, mametest
+from tools.build import mkcpmfs, mametest, mkcbmfs
 from build.llvm import llvmrawprogram, llvmclibrary
 from config import (
     MINIMAL_APPS,
@@ -19,42 +19,6 @@ COMMODORE_ITEMS = (
 )
 
 COMMODORE_ITEMS_WITH_SCREEN = COMMODORE_ITEMS | SCREEN_APPS | SCREEN_APPS_SRCS
-
-
-@Rule
-def mkcbmfs(
-    self, name, items: TargetsMap = {}, type="d64", title="CBMFS", id=""
-):
-    cs = []
-    ins = []
-    deps = ["tools+mkcombifs"]
-
-    if type == "d2m":
-        cs += [f"zcat $[deps[1]] > $[outs[0]]"]
-        deps += ["extras/empty.d2m.gz"]
-    else:
-        cs += [f"chronic c1541 -format '{title}',{id} {type} $[outs[0]]"]
-
-    cmd = f"chronic c1541 $[outs[0]] -name '{title}'"
-    for k, v in items.items():
-        t = "p"
-        if k.startswith("&"):
-            t = "u"
-
-        cmd += f" -write '{filenameof(v)}' '{k},{t}'"
-        ins += [v]
-    cs += [cmd]
-
-    if type != "d2m":
-        cs += ["$[deps[0]] -f $[outs[0]]"]
-    simplerule(
-        replaces=self,
-        ins=ins,
-        outs=[f"={name}.{type}"],
-        deps=deps,
-        commands=cs,
-        label="MKCBMFS",
-    )
 
 
 @Rule
